parsers/queuejobs/requirements.txt
azure-functions
requests==2.21.0
azure-storage-queue==12.1.2